# api/openapi.yaml
openapi: 3.0.3
info:
  title: EgressDeck Controller HTTP API
  version: 0.1.0
  description: Current Go HTTP contract. Snake_case wire values are mapped by the Vue client into UI summaries.
    Collection pages use items, limit, total and optional next_cursor; the durable event feed uses its own resume
    cursor. Conditional remote adapter routes return typed HTTP 501 when unavailable. A 202 identifies an existing
    controller journal entry and is separate from applied, observed and traffic-verified state. Roadmap-only operations
    are outside paths under x-planned-operations; OIDC callback registration follows exact OIDC_REDIRECT_URL under
    x-oidc-callback.
servers:
- url: /
security:
- sessionCookie: []
...
paths:
...
  /api/v1/capabilities:
    get:
      tags:
      - gateways
      summary: Read configured controller capabilities
      operationId: getCapabilities
      responses:
        '200':
          description: Capability flags
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ControllerCapabilities'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-implementation-status: implemented
      description: Controller flags describe configured service availability. events.resume=true reflects the local
        persisted polling feed. Diagnostics still read live gateway capabilities on each request. policy.apply_generation
        is an explicit qualification flag; the presence of an executor alone does not enable it. Provider/selection
        flags do not establish real-network acceptance.
      x-required-role: viewer
  /api/v1/overview:
    get:
      tags:
      - inventory
      summary: Read controller overview
      operationId: getOverview
      responses:
        '200':
          description: Overview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Overview'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-implementation-status: implemented
      description: Returns provider lifecycle and cached firewall projections. Missing firewall counts/drift mean
        unknown; observed_at and observation_status distinguish fresh, stale, error or never_read. coverage is null
        with coverage_status=unavailable.
      x-required-role: viewer
  /api/v1/devices:
    get:
      tags:
      - inventory
      summary: List devices
      operationId: listDevices
      responses:
        '200':
          description: Collection in items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceList'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: invalid_cursor or invalid query/limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'stale_cursor: snapshot changed or cursor expired'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: collection_item_too_large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: collection_capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Returns a bounded items page with limit, total and optional next_cursor. Default limit 100, maximum
        500. Follow next_cursor using the same query and limit. Snapshot change or expiry returns 409; restart invalidates
        the cursor. Responses are bounded to 4 MiB, scan to 100000 items/64 MiB; live observation values can refresh
        between pages.
      parameters:
      - $ref: '#/components/parameters/CollectionLimit'
      - $ref: '#/components/parameters/CollectionCursor'
      x-required-role: viewer
    post:
      tags:
      - inventory
      summary: Create unmanaged device intent
      operationId: createDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '201':
          description: Device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/CSRF'
      x-required-role: admin
  /api/v1/devices/{id}:
    parameters:
    - $ref: '#/components/parameters/ResourceID'
    get:
      tags:
      - inventory
      summary: Read a device
      operationId: getDevice
      responses:
        '200':
          description: Device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          $ref: '#/components/responses/Error404'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
      x-implementation-status: implemented
      x-required-role: viewer
    put:
      tags:
      - inventory
      summary: Replace device intent
      operationId: replaceDevice
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/Revision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '200':
          description: Device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '428':
          description: Expected revision header/query required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Replaces the entire resource; PATCH is not JSON merge patch. Supply If-Match or revision query.
        Omitted mutable fields use zero values.
      x-required-role: admin
    patch:
      tags:
      - inventory
      summary: Patch device intent
      operationId: patchDevice
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/Revision'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceCreate'
      responses:
        '200':
          description: Device
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '412':
          $ref: '#/components/responses/Error412'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '428':
          description: Expected revision header/query required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Replaces the entire resource; PATCH is not JSON merge patch. Supply If-Match or revision query.
        Omitted mutable fields use zero values.
      x-required-role: admin
    delete:
      tags:
      - inventory
      summary: Delete an unenrolled device
      operationId: deleteDevice
      parameters:
      - $ref: '#/components/parameters/IfMatch'
      - $ref: '#/components/parameters/CSRF'
      - $ref: '#/components/parameters/Revision'
      responses:
        '204':
          description: Deleted
        '409':
          $ref: '#/components/responses/Error409'
        '412':
          $ref: '#/components/responses/Error412'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '428':
          description: Expected revision header/query required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      x-required-role: admin
  /api/v1/device-groups:
    get:
      tags:
      - inventory
      summary: List device groups
      operationId: listDeviceGroups
      responses:
        '200':
          description: Collection in items
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroupList'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: invalid_cursor or invalid query/limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'stale_cursor: snapshot changed or cursor expired'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: collection_item_too_large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: collection_capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Returns a bounded items page with limit, total and optional next_cursor. Default limit 100, maximum
        500. Follow next_cursor using the same query and limit. Snapshot change or expiry returns 409; restart invalidates
        the cursor. Responses are bounded to 4 MiB, scan to 100000 items/64 MiB; live observation values can refresh
        between pages.
      parameters:
      - $ref: '#/components/parameters/CollectionLimit'
      - $ref: '#/components/parameters/CollectionCursor'
      x-required-role: viewer
    post:
      tags:
      - inventory
      summary: Create a device group
      operationId: createDeviceGroup
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DeviceGroupCreate'
      responses:
        '201':
          description: Device group
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceGroup'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/CSRF'
      x-required-role: admin
...
  /api/v1/policies:
    get:
      tags:
      - policy
      summary: List typed policies
      operationId: listPolicies
      responses:
        '200':
          description: Policies
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolicyList'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: invalid_cursor or invalid query/limit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: 'stale_cursor: snapshot changed or cursor expired'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: collection_item_too_large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: collection_capacity
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      description: Returns a bounded items page with limit, total and optional next_cursor. Default limit 100, maximum
        500. Follow next_cursor using the same query and limit. Snapshot change or expiry returns 409; restart invalidates
        the cursor. Responses are bounded to 4 MiB, scan to 100000 items/64 MiB; live observation values can refresh
        between pages.
      parameters:
      - $ref: '#/components/parameters/CollectionLimit'
      - $ref: '#/components/parameters/CollectionCursor'
      x-required-role: viewer
    post:
      tags:
      - policy
      summary: Create a typed policy
      operationId: createPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PolicyWrite'
      responses:
        '201':
          description: Policy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Policy'
        '409':
          $ref: '#/components/responses/Error409'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/CSRF'
      x-required-role: admin
...
  /api/v1/policies/explain:
    post:
      tags:
      - policy
      summary: Explain a predicted policy match
      operationId: explainPolicy
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExplainRequest'
      responses:
        '200':
          description: Predicted explanation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Explanation'
        '422':
          $ref: '#/components/responses/Error422'
        '401':
          $ref: '#/components/responses/Error401'
        '403':
          $ref: '#/components/responses/Error403'
        '400':
          description: Invalid JSON, unknown fields or request validation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      x-implementation-status: implemented
      parameters:
      - $ref: '#/components/parameters/CSRF'
      x-required-role: operator
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: egressdeck_session
      description: HTTP-only session; mutations also require X-CSRF-Token matching egressdeck_csrf.
...
  parameters:
...
    CollectionLimit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 100
...
  responses:
...
    Error401:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Error403:
      description: Forbidden or CSRF failure
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
...
    ControllerCapabilities:
      type: object
      required:
      - api_version
      - capabilities
      properties:
        api_version:
          type: string
        capabilities:
          type: object
          additionalProperties:
            type: boolean
...
    DeviceList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        limit:
          type: integer
          minimum: 1
          maximum: 500
        total:
          type: integer
          minimum: 0
        next_cursor:
          type: string
      additionalProperties: false
      required:
      - items
      - limit
      - total
...